fix: never exclude Dockerfile
I just noticed builds started failing - I have `Dockerfile` in `.dockerignore`. 

It makes sense for the purpose of the `.dockerignore` file but it broke builds here. There would be no reason to ever exclude it for this construct so I added a final pass to remove it from the exclusion list if present.